\name{FCSGenerator2.5.run}
\alias{FCSGenerator2.5.run}
\title{FCS Generator 2.5 - Main Function}
\usage{
FCSGenerator2.5.run()
}
\description{
Launches the shiny app FCS Generator 2.5. FCS Generator 2.5 is a tool used you to generate FCS files from scratch or use cytometry data to generate a set of "ground truth" files meant to benchmark algrotihms such as clustering ones, or dimensionality reduction methods.
}
\examples{
FCSGenerator2.5.run()
}
