\name{FCSGenerator2.run}
\alias{FCSGenerator2.run}
\title{FCS Generator 2 - Main Function}
\usage{
FCSGenerator2.run()
}
\description{
Launches the shiny app FCS Generator2. FCS Generator2 is a tool allowing you to generate FCS files from scratch or use cytometry data to generate a set of "ground truth" files meant to benchmark algrotihms such as clustering ones, or dimensionality reduction methods.
}
\examples{
FCSGenerator2.run()
}
